# Documentation/http-api/openapi.json
openapi: "3.0.0"
info:
  version: 1.0.0
  title: COI Push Service
  license:
    name: GNU GENERAL PUBLIC LICENSE
servers:
  - url: "http://localhost:8123/"
    description: "For testing against localhost"
paths:
...
  /invite/:
    put:
      summary: Creates a new invitation
      operationId: createInvitation
      tags:
        - invitation
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/Invitation"
      responses:
        '201':
          description: The created invitation
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/Invitation"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /invite/{id}:
    get:
      summary: Gets an invitation
      operationId: getInvitation
      tags:
        - invitation
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the invitation
          schema:
            type: string
      responses:
        '200':
          description: The invitation
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/Invitation"
        '404':
          description: An invitation with this id doesn't exist
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Deletes an invitation
      operationId: deleteInvitation
      tags:
        - invitation
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the invitation
          schema:
            type: string     
      responses:
        '200':
          description: An empty response in case of success
        '404':
          description: An invitation with this id doesn't exist
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"                
components:
  schemas:
...
    Invitation:
      properties:
        message:
          description: The message of the invitation
          type: string
        endpoint:
          description: The endpoint for receiving the invitation
          type: string
        id:
          description: The id of the invitation
          type: string
        sender:
          description: The sender of the invitation
          type: object
          properties:
            email:
              description: The email of the sender
              type: string
            name:
              description: The name of the sender
              type: string
            image:
              description: The image of the sender as a base64 encoded string
              type: string
            publicKey:
              description: The public key of the sender
              type: string      
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
  securitySchemes:
    vapid:
      type: http
      scheme: vapid
      description: See VAPID RFC 8292